pyproject.toml
[tool.poetry]
name = "smart-resume-builder-backend"
version = "0.1.0"
description = "Backend for Smart Resume Builder and Job Searcher"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1"
djangorestframework = "^3.15"
django-cors-headers = "^4.4"
psycopg = {extras = ["binary"], version = "^3.2"}
dj-database-url = "^2.2"
python-dotenv = "^1.0"
supabase = "^2.8"
openai = "^1.51"
langchain = "^0.3"
langchain-openai = "^0.2"
langchain-community = "^0.3"
pydantic = "^2.9"
celery = {extras = ["redis"], version = "^5.4"}
redis = "^5.0"
gunicorn = "^23.0"
whitenoise = "^6.7"
pinecone-client = "^5.0"
pypdf = "^5.0"
python-docx = "^1.1"
# unstructured = {version = "^0.15", extras = ["all-docs"]}  # Note: Requires Python <3.13, comment out if using Python 3.13+
tiktoken = "^0.8"
requests = "^2.32"
pillow = "^10.4"
drf-spectacular = "^0.27"
sentry-sdk = {extras = ["django"], version = "^2.14"}
weasyprint = "^66.0"
qrcode = {version = "^7.4", extras = ["pil"]}
playwright = "^1.40.0"
reportlab = "^4.4.5"
PyMuPDF = "^1.23.0"
textstat = "^0.7.3"
rapidfuzz = "^3.9"
nltk = "^3.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-django = "^4.7"
pytest-cov = "^4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

